0.7.0 - E nhance documentation for event handling, and refactor card-related imports for consistency. Remove deprecated Card class and related methods, streamlining the codebase.
