+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "shoehive",
-  "version": "0.6.0",
+  "version": "0.7.0",
   "description": "WebSocket-based multiplayer game framework for real-time, event-driven gameplay",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
