+++ tests/core/TableUncovered.test.ts
@@ -1,7 +1,7 @@
 import { Table, TableState } from '../../src/core/Table';
 import { EventBus } from '../../src/events/EventBus';
 import { Player } from '../../src/core/Player';
-import { Hand, Card, CardSuit, CardRank, Deck } from '../../src/core/Card';
+import { Hand, Card, CardSuit, CardRank, Deck } from '../../src/core/card';
 import * as WebSocket from 'ws';
 import { TABLE_EVENTS } from '../../src/events/EventTypes';
